2.2.3.47 changes
Note: This version might remove some aluminium-related items.

Added Magneticraft module
Added feature to entry types where one can make new types
Added ore rarities
Added remaps
Added oredict compat for ore name nonsense

Code cleanup
Changed all modules' internal maps to use ore entries as keys since they
should never change
Made sure minecraft ore stuff more preferred, and JAOPCA ore stuff least
preferred

+++ src/main/java/thelm/jaopca/api/block/BlockBase.java
@@ -254,7 +254,7 @@ public class BlockBase extends Block implements IBlockWithProperty {
 
 			if(!BlockFalling.fallInstantly && worldIn.isAreaLoaded(pos.add(-32, -32, -32), pos.add(32, 32, 32))) {
 				if(!worldIn.isRemote) {
-					EntityFallingBlock entityfallingblock = new EntityFallingBlock(worldIn, (double)pos.getX() + 0.5D, (double)pos.getY(), (double)pos.getZ() + 0.5D, worldIn.getBlockState(pos));
+					EntityFallingBlock entityfallingblock = new EntityFallingBlock(worldIn, pos.getX() + 0.5D, pos.getY(), pos.getZ() + 0.5D, worldIn.getBlockState(pos));
 					worldIn.spawnEntity(entityfallingblock);
 				}
 			}
@@ -285,9 +285,9 @@ public class BlockBase extends Block implements IBlockWithProperty {
 			BlockPos blockpos = pos.down();
 
 			if(BlockFalling.canFallThrough(worldIn.getBlockState(blockpos))) {
-				double d0 = (double)((float)pos.getX() + rand.nextFloat());
-				double d1 = (double)pos.getY() - 0.05D;
-				double d2 = (double)((float)pos.getZ() + rand.nextFloat());
+				double d0 = pos.getX() + rand.nextFloat();
+				double d1 = pos.getY() - 0.05D;
+				double d2 = pos.getZ() + rand.nextFloat();
 				worldIn.spawnParticle(EnumParticleTypes.FALLING_DUST, d0, d1, d2, 0.0D, 0.0D, 0.0D, new int[] {Block.getStateId(stateIn)});
 			}
 		}
